backend/services/marker_tool/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
distlib==0.4.0
distro==1.9.0
einops==0.8.1
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
ftfy==6.3.1
google-auth==2.41.1
google-genai==1.39.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
identify==2.6.14
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
markdown2==2.5.4
markdownify==1.2.0
marker-pdf==1.10.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
opencv-python-headless==4.11.0.86
packaging==25.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.11.0
PyMuPDF==1.26.4
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2024.11.6
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
surya-ocr==0.17.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
websockets==15.0.1

